backend/requirements-test.txt
# requirements-test.txt - Testing dependencies for Wasteland Tarot API

# Core Testing Framework
pytest>=7.4.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0
pytest-timeout>=2.2.0
pytest-xdist>=3.5.0  # Parallel test execution

# HTTP Testing
httpx>=0.26.0
pytest-httpx>=0.27.0

# Database Testing
pytest-postgresql>=5.0.0
pytest-alembic>=0.10.0
factory-boy>=3.3.0
sqlalchemy-utils>=0.41.0

# Data Generation and Mocking
faker>=21.0.0
freezegun>=1.4.0
responses>=0.24.0

# Performance Testing
pytest-benchmark>=4.0.0
psutil>=5.9.0
memory-profiler>=0.61.0

# Redis Testing (if using Redis)
pytest-redis>=3.0.0
fakeredis>=2.20.0

# Security Testing
pytest-security>=0.1.0
bandit>=1.7.5

# Code Quality
black>=23.12.0
isort>=5.13.0
flake8>=6.1.0
mypy>=1.8.0
pylint>=3.0.0

# Documentation Testing
pytest-doctestplus>=1.0.0

# API Documentation Testing
openapi-spec-validator>=0.7.0

# Audio Testing (for Fallout audio features)
pydub>=0.25.1
mutagen>=1.47.0

# Reporting
pytest-html>=4.1.0
pytest-json-report>=1.5.0
allure-pytest>=2.13.0

# Development Utilities
ipdb>=0.13.13
pytest-watch>=4.2.0
pytest-sugar>=0.9.7  # Better test output formatting

# Environment Management
python-dotenv>=1.0.0
pydantic-settings>=2.1.0

# Additional Async Testing
anyio>=4.2.0
trio>=0.24.0

# Load Testing
locust>=2.17.0

# Snapshot Testing
syrupy>=4.6.0

# Property-based Testing
hypothesis>=6.92.0

# Test Data Management
mimesis>=15.1.0

# JSON Schema Validation
jsonschema>=4.20.0

# Time Zone Testing
pytz>=2023.3